Allow saving streams to gzipped csv or SQLite
I've included two functions that take the stream iterator and save the
incoming items to a file. The SQLite file allows easy review and analysis
of the data. The main advantage of the gzipped CSV format is that it
produces a smaller file. It has the same semi-parsed information with
each row representing an item in the stream. This should allow it to be
quickly read into a database for analysis.